Extraido el FakeMovieFactory del TestMoviesLocalSource
